1. configure mvc default servlet handler for static resources 2. do filter with transaction scope to the request servlet path is starts with '/fx/' 3. fix typo

+++ src/main/java/org/silentsoft/everywhere/server/core/filter/AuthenticationFilter.java
@@ -14,6 +14,7 @@ import javax.servlet.http.HttpServletResponse;
 import org.silentsoft.core.util.ObjectUtil;
 import org.silentsoft.everywhere.context.BizConst;
 import org.silentsoft.everywhere.context.util.SecurityUtil;
+import org.silentsoft.everywhere.server.util.TransactionUtil;
 import org.silentsoft.io.memory.SharedThreadMemory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -30,29 +31,52 @@ public class AuthenticationFilter implements Filter {
 		
 		HttpServletRequest request = (HttpServletRequest) servletRequest;
 
-		if ("/fx/register/authentication".equals(request.getServletPath()) ||
-			"/fx/login/authentication".equals(request.getServletPath())) {
-			filterChain.doFilter(servletRequest, servletResponse);
-		} else {
-			if (isValidUser(request)) {
-				SharedThreadMemory.create();
-				
-				SharedThreadMemory.put(BizConst.KEY_USER_ID, request.getHeader("user"));
-				SharedThreadMemory.put(BizConst.KEY_USER_UNIQUE_SEQ, request.getHeader("sequence"));
-				filterChain.doFilter(servletRequest, servletResponse);
-				
-				SharedThreadMemory.delete();
+		/**
+		 * if the path is starts with "/fx/" then must be filter with transaction scope.
+		 * the request that path starts with fx is triggered by client side application.
+		 */
+		if (request.getServletPath().length() > 0 && request.getServletPath().startsWith("/fx/")) {
+			if ("/fx/register/authentication".equals(request.getServletPath()) ||
+				"/fx/login/authentication".equals(request.getServletPath())) {
+				doFilterWithTransaction(filterChain, servletRequest, servletResponse);
 			} else {
-				LOGGER.info("Detected not valid user on <{}> from <{}> !", new Object[]{request.getRequestURI(), request.getRemoteAddr()});
-				
-				HttpServletResponse response = (HttpServletResponse) servletResponse;
-				response.sendError(HttpServletResponse.SC_BAD_REQUEST, "[WARNING] You are trying to access server by unknown way !!!");
+				if (isValidUser(request)) {
+					SharedThreadMemory.create();
+					
+					SharedThreadMemory.put(BizConst.KEY_USER_ID, request.getHeader("user"));
+					SharedThreadMemory.put(BizConst.KEY_USER_UNIQUE_SEQ, request.getHeader("sequence"));
+					
+					doFilterWithTransaction(filterChain, servletRequest, servletResponse);
+					
+					SharedThreadMemory.delete();
+				} else {
+					LOGGER.info("Detected not valid user on <{}> from <{}> !", new Object[]{request.getRequestURI(), request.getRemoteAddr()});
+					
+					HttpServletResponse response = (HttpServletResponse) servletResponse;
+					response.sendError(HttpServletResponse.SC_BAD_REQUEST, "[WARNING] You are trying to access server by unknown way !!!");
+				}
 			}
+		} else {
+			// not really triggered by c/s but static resources.
+			filterChain.doFilter(servletRequest, servletResponse);
 		}
 	}
 	
 	public void destroy() {
 		
+	}
+	
+	private boolean doFilterWithTransaction(FilterChain filterChain, ServletRequest servletRequest, ServletResponse servletResponse) throws IOException, ServletException {
+		return TransactionUtil.doScope(
+			() -> {
+				filterChain.doFilter(servletRequest, servletResponse);
+				
+				return true;
+			},
+			() -> {
+				return false;
+			}
+		);
 	}
 
 	private boolean isValidUser(HttpServletRequest request) {
